test: Combine NATS and ETCD start/stop into conftest.py
