+++ lib/bindings/python/tests/test_kv_bindings.py
@@ -15,8 +15,6 @@
 
 
 import asyncio
-import subprocess
-from time import sleep
 from typing import List
 
 import pytest
@@ -35,24 +33,6 @@ from dynamo.llm import (
 from dynamo.runtime import Component, DistributedRuntime
 
 pytestmark = pytest.mark.pre_merge
-
-
-@pytest.fixture(scope="module", autouse=True)
-def setup_and_teardown():
-    # Setup code
-    nats_server = subprocess.Popen(["nats-server", "-js"])
-    etcd = subprocess.Popen(["etcd"])
-    print("Setting up resources")
-
-    sleep(5)  # wait for nats-server and etcd to start
-    yield
-
-    # Teardown code
-    print("Tearing down resources")
-    nats_server.terminate()
-    nats_server.wait()
-    etcd.terminate()
-    etcd.wait()
 
 
 @pytest.fixture(scope="module")
